Add common tools module
