--- examples/post_training_quantization/onnx/mobilenet_v2/__init__.py
@@ -1,1 +1,0 @@
-../../../__init__.py
