+++ examples/post_training_quantization/onnx/mobilenet_v2/common.py
@@ -1,0 +1,1 @@
+../../../common.py
